+++ console2/src/components/molecules/ProcessList/index.tsx
@@ -42,6 +42,12 @@ export const PROJECT_COLUMN: ColumnDefinition = {
     render: 'project-link'
 };
 
+export const REPO_COLUMN: ColumnDefinition = {
+    caption: 'Repository',
+    source: 'repoName',
+    render: 'repo-link'
+}
+
 export const INITIATOR_COLUMN: ColumnDefinition = {
     caption: 'Initiator',
     source: 'initiator'
@@ -100,6 +106,8 @@ const renderColumnContent = (e: ProcessEntry, c: ColumnDefinition) => {
         return <LocalTimestamp value={v} />;
     } else if (c.render === 'project-link') {
         return <Link to={`/org/${e.orgName}/project/${e.projectName}`}>{v}</Link>;
+    } else if (c.render === 'repo-link') {
+        return <Link to={`/org/${e.orgName}/project/${e.projectName}/repository/${e.repoName}`}>{v}</Link>;
     }
 
     return v;